separate policer-list
