+++ catalystwan/endpoints/configuration/policy/list/policer.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import PolicerList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.policer import PolicerList, PolicerListEditPayload, PolicerListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class PolicerListEditPayload(PolicerList, PolicyListId):
-    pass
-
-
-class PolicerListInfo(PolicerList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyPolicerClassList(APIEndpoints, PolicyListEndpoints):
